requirements.txt
azure-functions
requests
meteostat
geopy